+++ is-ios-firefox.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = function isIosFirefox(ua) {
+  ua = ua || global.navigator.userAgent;
+  return /FxiOS/i.test(ua);
+};
+