+++ supports-popups.js
@@ -3,6 +3,7 @@
 var MINIMUM_SUPPORTED_CHROME_IOS_VERSION = 48;
 
 var isAndroid = require('./is-android');
+var isIosFirefox = require('./is-ios-firefox');
 var isIosWebview = require('./is-ios-webview');
 var isChrome = require('./is-chrome');
 var isSamsungBrowser = require('./is-samsung');
@@ -43,5 +44,5 @@ function isOldSamsungBrowserOrSamsungWebview(ua) {
 
 module.exports = function supportsPopups(ua) {
   ua = ua || global.navigator.userAgent;
-  return !(isIosWebview(ua) || isAndroidWebview(ua) || isOperaMini(ua) || isUnsupportedIosChrome(ua) || isOldSamsungBrowserOrSamsungWebview(ua));
+  return !(isIosWebview(ua) || isIosFirefox(ua) || isAndroidWebview(ua) || isOperaMini(ua) || isUnsupportedIosChrome(ua) || isOldSamsungBrowserOrSamsungWebview(ua));
 };