+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## Unreleased
 
+* Correct is-chrome to not return true on edge browser
+* Correct supports-popups to not return true on iOS Firefox
 * Add isSamsungBrowser
 * Add isMobileFirefox
+* Add isIosFirefox
 
 ## 1.5.0 (2017-07-03)